introduced AbstractField, started randomDestroyerBot
AbstractField is used by fieldMemory in AbstractPlayer to store the
whole field as it gets uncovered. Started logic of randomDestroyer bot
with a goal-object that takes control from the first hit of a ship
until it is destroyed. Not sure yet what’s the best way to go about it.